backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-cloud-bigquery>=3.14.0
db-dtypes>=1.1.0
pandas>=2.1.0
numpy>=1.26.0
cachetools>=5.3.0
python-dateutil>=2.8.0
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
